Validierung Prüfung editieren/löschen Zeitlogik

+++ src/main/java/de/nordakademie/multiplechoice/exam/ui/ExamDetailAction.java
@@ -14,6 +14,8 @@ import de.nordakademie.multiplechoice.user.service.UserService;
 import org.apache.struts2.interceptor.SessionAware;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.sql.Date;
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Map;
 
@@ -74,27 +76,49 @@ public class ExamDetailAction extends ActionSupport implements Action, SessionAw
     }
 
     public String removeExam() {
-        questions = questionService.findByExam(examId);
-        for (Question question : questions) {
-            answers = answerService.findAll(question.getId());
-            for (Answer answer : answers) {
-                answerService.delete(answer.getAnswerID());
+        exam = examService.findOne(examId);
+
+        if (!isEditableExam(examId, exam.getEnd())) {
+            questions = questionService.findByExam(examId);
+            for (Question question : questions) {
+                answers = answerService.findAll(question.getId());
+                for (Answer answer : answers) {
+                    answerService.delete(answer.getAnswerID());
+                }
+                questionService.delete(question.getId());
             }
-            questionService.delete(question.getId());
+            participationService.deleteExam(examId);
+            examService.removeExam(this.getExamId());
+            return SUCCESS;
+        } else {
+            addActionError("Prüfung kann erst nach Ende des Prüfungszeitraums nicht mehr gelöscht werden");
+            return INPUT;
         }
-        participationService.deleteExam(examId);
-        examService.removeExam(this.getExamId());
-        return SUCCESS;
     }
 
-    public String viewExam() {
-
+    private void loadExamData() {
         participations = participationService.findAll(this.getExamId());
         exam = examService.findOne(this.getExamId());
         questions = questionService.findByExam(exam.getId());
+    }
 
+    public String editExam() {
+        exam = examService.findOne(examId);
+
+        if (isEditableExam(examId, exam.getStart())) {
+            loadExamData();
+            return SUCCESS;
+        } else {
+            addActionError("Prüfung kann nach Beginn des Prüfungszeitraums nicht mehr editiert werden");
+            return INPUT;
+        }
+    }
+
+    public String viewExam() {
+        loadExamData();
         return SUCCESS;
     }
+
 
     public String saveExam() {
 
@@ -110,11 +134,17 @@ public class ExamDetailAction extends ActionSupport implements Action, SessionAw
         return SUCCESS;
     }
 
-    public void validate(){
+    public void validate() {
     }
 
-    public boolean isEditableExam() {
-        return editableExam;
+
+    private boolean isEditableExam(long examId, Date date) {
+        Timestamp currentTimestamp = new Timestamp(System.currentTimeMillis());
+        if (date.before(currentTimestamp)) {
+            return false;
+        } else {
+            return true;
+        }
     }
 
     public void setEditableExam(boolean editableExam) {
